configuração do servidor

+++ src/Database/Entities/Pizza.ts
@@ -1,4 +1,4 @@
-import { Column, Entity, PrimaryGeneratedColumn } from "typeorm";
+import { Column, Entity, BaseEntity, PrimaryGeneratedColumn } from "typeorm";
 import IPizza from "../../Interfaces/IPizza";
 
 @Entity("pizzas")
@@ -12,7 +12,8 @@ export default class Pizza implements IPizza {
   @Column({ type: "text" })
   type: string;
 
-  @Column({ type: "int" })
+
+  @Column({ type: "decimal" })
   price: number;
 
   @Column({ type: "text" })
